Minor bug tsx -> ts
